Use react router hooks to fetch and move to details
