+++ src/state/cardDetails/services/index.js
@@ -1,3 +1,4 @@
 import { fakeHttpModule } from "common/http";
 
-export const getDetails = ({payload: {url}}) => fakeHttpModule.get(url, process.env.REACT_APP_TOKEN);
+export const getDetails = ({ payload: { url } }) =>
+	fakeHttpModule.get(url /* process.env.REACT_APP_TOKEN */);
